CAMEL-17018: rework the Camel JBang code to avoid build-time version alignment
This should prevent Camel JBang from aligning to snapshot versions of
the dependencies during release builds.
