+++ dsl/camel-jbang/camel-jbang-main/pom.xml
@@ -42,58 +42,8 @@
     <dependencies>
         <dependency>
             <groupId>org.apache.camel</groupId>
-            <artifactId>camel-core</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>org.apache.camel</groupId>
-            <artifactId>camel-core-model</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>org.apache.camel</groupId>
-            <artifactId>camel-api</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>org.apache.camel</groupId>
-            <artifactId>camel-main</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>org.apache.camel</groupId>
-            <artifactId>camel-kamelet-main</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>org.apache.camel</groupId>
-            <artifactId>camel-file-watch</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>org.apache.camel</groupId>
-            <artifactId>camel-resourceresolver-github</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>org.apache.camel</groupId>
             <artifactId>camel-jbang-core</artifactId>
             <version>${project.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>org.apache.logging.log4j</groupId>
-            <artifactId>log4j-api</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>org.apache.logging.log4j</groupId>
-            <artifactId>log4j-core</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>org.apache.logging.log4j</groupId>
-            <artifactId>log4j-slf4j-impl</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>info.picocli</groupId>
-            <artifactId>picocli</artifactId>
-        </dependency>
-
-        <dependency>
-            <groupId>org.apache.velocity</groupId>
-            <artifactId>velocity-engine-core</artifactId>
-            <version>${velocity-version}</version>
         </dependency>
     </dependencies>
 
@@ -114,6 +64,7 @@
                 <artifactId>maven-checkstyle-plugin</artifactId>
                 <configuration>
                     <skip>${checkstyle.skip}</skip>
+                    <failOnViolation>false</failOnViolation>
                 </configuration>
             </plugin>
 
@@ -124,15 +75,11 @@
                         <phase>prepare-package</phase>
                         <configuration>
                             <tasks>
-                                <replace token="${camel.jbang.log4j2.version}" value="${log4j2-version}" dir="${dist.dir}">
-                                    <include name="**/*.java"/>
-                                </replace>
-                                <replace token="${camel.jbang.picocli.version}" value="${picocli-version}" dir="${dist.dir}">
-                                    <include name="**/*.java"/>
-                                </replace>
-                                <replace token="${camel.jbang.velocity.version}" value="${velocity-version}" dir="${dist.dir}">
-                                    <include name="**/*.java"/>
-                                </replace>
+                                <copy todir="${dist.dir}">
+                                    <fileset dir="${project.build.directory}">
+                                        <include name="**/*.java"/>
+                                    </fileset>
+                                </copy>
                                 <chmod file="${dist.dir}/*.java" perm="u+x" />
                             </tasks>
                         </configuration>
